pyproject.toml
[tool.poetry]
name = "meguchan-lte"
version = "0.1.0"
description = "meguchan-lte"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.5"
nonebot2 = "^2.0.0-beta.1"
nonebot-adapter-onebot = "^2.0.0-beta.1"
nonebot-plugin-gocqhttp = "^0.4.6"

emoji = "^1.7.0"
aiohttp = "^3.8.1"
mcstatus = "^9.2.0"
aio-mc-rcon = "^3.1.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = [
    "nonebot_plugin_gocqhttp",
    "nonebot_plugin_apscheduler",
    "nonebot_plugin_help"
]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
